requirements.txt
certifi>=2024.0.0
python-dotenv>=1.0.0
